Another restructuration of the project
