+++ core/codec.go
@@ -3,13 +3,15 @@ package core
 import (
 	"fmt"
 	"reflect"
+
+    "essai/ntfstool/core/dataio"
 )
 
 type IEncoder interface {
 	Encode(interface{}) error
 }
 
-func WriteRecord(encoder IEncoder, record IDataRecord) error {
+func WriteRecord(encoder IEncoder, record dataio.IDataRecord) error {
 	return encoder.Encode(record)
 }
 
@@ -17,13 +19,13 @@ type IDecoder interface {
 	Decode() (interface{}, error)
 }
 
-func ReadRecord(decoder IDecoder) (IDataRecord, error) {
+func ReadRecord(decoder IDecoder) (dataio.IDataRecord, error) {
 	val, err := decoder.Decode()
 	if err != nil {
 		return nil, err
 	}
 
-	res, ok := val.(IDataRecord)
+	res, ok := val.(dataio.IDataRecord)
 	if !ok {
 		return nil, WrapError(fmt.Errorf("`%s` is not a data record", reflect.TypeOf(val)))
 	}
