+++ core/dataio/codec/entry.go
