+++ core/dataio/datafile/datafile.go
@@ -1,0 +1,6 @@
+package datafile
+
+type tIndex struct {
+	start  int64
+	length int64
+}
