+++ core/dataio/dataio.go
@@ -1,4 +1,4 @@
-package defs
+package dataio
 
 import "fmt"
